pyproject.toml
[project]
name = "dataxtr"
version = "0.1.0"
description = "Agentic data extraction system using LangGraph"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["langgraph", "data-extraction", "llm", "agents", "document-processing"]

dependencies = [
    # Core LangChain/LangGraph
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    # LLM Providers
    "langchain-anthropic>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-google-genai>=2.0.0",
    # Document Processing
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "pandas>=2.2.0",
    # OCR and Vision
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
    "pdf2image>=1.17.0",
    # Table Extraction
    "camelot-py>=0.11.0",
    # Data Validation
    "pydantic>=2.5.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
    "aiofiles>=24.1.0",
    "langchain-groq>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dataxtr"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
